made CLoudQuery class and returned its object in response
